# libs/shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Web CAT API
  version: 0.1.0
  description: Initial API contract for the Web CAT platform.
servers:
  - url: http://localhost:8000
paths:
...
  /documents:
    get:
      summary: List imported documents
      operationId: listDocuments
      tags:
        - documents
      responses:
        "200":
          description: Imported documents ordered by creation date.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      summary: Import a TXT document
      operationId: importTxtDocument
      tags:
        - documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentImportRequest"
      responses:
        "201":
          description: Imported document with generated segments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "400":
          description: Text could not be segmented.
        "404":
          description: Project was not found.
  /documents/{document_id}:
    get:
      summary: Get document with segments
      operationId: getDocument
      tags:
        - documents
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Document details and segment list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetail"
        "404":
          description: Document was not found.
  /documents/{document_id}/segments:
    get:
      summary: List document segments
      operationId: listDocumentSegments
      tags:
        - documents
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Segments ordered by source position.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Segment"
        "404":
          description: Document was not found.
  /documents/{document_id}/export.txt:
    get:
      summary: Export translated document as TXT
      operationId: exportTxtDocument
      tags:
        - documents
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Plain text document ordered by segment position. Missing targets fall back to source text.
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Document was not found.
  /documents/{document_id}/export.xliff:
    get:
      summary: Export working document as XLIFF
      operationId: exportXliffDocument
      tags:
        - documents
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Minimal XLIFF 1.2 document with source, target, status and languages.
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            application/x-xliff+xml:
              schema:
                type: string
        "404":
          description: Document was not found.
...
  /glossary/search:
    post:
      summary: Search glossary terms in source text
      operationId: searchGlossaryTerms
      tags:
        - glossary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlossarySearchRequest"
      responses:
        "200":
          description: Glossary terms found in the submitted source text.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlossarySearchResponse"
  /glossary/terms:
    get:
      summary: List glossary terms
      operationId: listGlossaryTerms
      tags:
        - glossary
      parameters:
        - name: source_language
          in: query
          required: false
          schema:
            type: string
        - name: target_language
          in: query
          required: false
          schema:
            type: string
        - name: domain
          in: query
          required: false
          schema:
            type: string
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Glossary terms ordered by source and target term.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GlossaryTerm"
    post:
      summary: Create a glossary term
      operationId: createGlossaryTerm
      tags:
        - glossary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlossaryTermCreateRequest"
      responses:
        "201":
          description: Created glossary term.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlossaryTerm"
  /glossary/terms/{term_id}:
    patch:
      summary: Update a glossary term
      operationId: updateGlossaryTerm
      tags:
        - glossary
      parameters:
        - name: term_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlossaryTermUpdateRequest"
      responses:
        "200":
          description: Updated glossary term.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlossaryTerm"
        "404":
          description: Glossary term was not found.
    delete:
      summary: Delete a glossary term
      operationId: deleteGlossaryTerm
      tags:
        - glossary
      parameters:
        - name: term_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Glossary term deleted.
        "404":
          description: Glossary term was not found.
  /glossary/import-csv:
    post:
      summary: Import glossary terms from CSV
      operationId: importGlossaryCsv
      tags:
        - glossary
      requestBody:
        required: true
        content:
          text/csv:
            schema:
              type: string
      responses:
        "201":
          description: Imported glossary terms.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlossaryImportResponse"
        "400":
          description: CSV header or rows are invalid.
  /glossary/export.tbx:
    get:
      summary: Export glossary as TBX
      operationId: exportGlossaryTbx
      tags:
        - glossary
      parameters:
        - name: source_language
          in: query
          required: false
          schema:
            type: string
        - name: target_language
          in: query
          required: false
          schema:
            type: string
        - name: domain
          in: query
          required: false
          schema:
            type: string
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: TBX export for matching glossary terms.
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            application/xml:
              schema:
                type: string
  /glossary/import-tbx:
    post:
      summary: Import glossary terms from TBX
      operationId: importGlossaryTbx
      tags:
        - glossary
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
      responses:
        "201":
          description: Imported or reused glossary terms.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlossaryImportResponse"
        "400":
          description: TBX XML or minimal structure is invalid.
...
components:
  schemas:
...
    Document:
      type: object
      required:
        - id
        - project_id
        - filename
        - source_language
        - target_language
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        filename:
          type: string
        source_language:
          type: string
        target_language:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    DocumentImportRequest:
      type: object
      required:
        - filename
        - content
      properties:
        filename:
          type: string
          minLength: 1
          maxLength: 255
          example: sample.txt
        content:
          type: string
          minLength: 1
          example: Save the file. Close the window.
        source_language:
          type: string
          minLength: 2
          maxLength: 16
          default: en
        target_language:
          type: string
          minLength: 2
          maxLength: 16
          default: pl
        project_id:
          type:
            - string
            - "null"
          format: uuid
        project_name:
          type:
            - string
            - "null"
          maxLength: 200
        segmentation_strategy:
          type: string
          enum:
            - sentence
            - paragraph
          default: sentence
    DocumentDetail:
      type: object
      required:
        - document
        - segments
      properties:
        document:
          $ref: "#/components/schemas/Document"
        segments:
          type: array
          items:
            $ref: "#/components/schemas/Segment"
    Segment:
      type: object
      required:
        - id
        - document_id
        - position
        - source_text
        - status
        - locked
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        position:
          type: integer
        source_text:
          type: string
        target_text:
          type:
            - string
            - "null"
        status:
          type: string
          enum:
            - new
            - draft
            - translated
            - reviewed
            - approved
        locked:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    GlossaryTermCreateRequest:
      type: object
      required:
        - source_language
        - target_language
        - source_term
        - target_term
      properties:
        project_id:
          type:
            - string
            - "null"
          format: uuid
        source_language:
          type: string
          minLength: 2
          maxLength: 16
        target_language:
          type: string
          minLength: 2
          maxLength: 16
        source_term:
          type: string
          minLength: 1
          maxLength: 255
        target_term:
          type: string
          minLength: 1
          maxLength: 255
        definition:
          type:
            - string
            - "null"
        domain:
          type:
            - string
            - "null"
          maxLength: 128
        case_sensitive:
          type: boolean
          default: false
        forbidden:
          type: boolean
          default: false
        example_source:
          type:
            - string
            - "null"
        example_target:
          type:
            - string
            - "null"
    GlossaryTermUpdateRequest:
      type: object
      properties:
        project_id:
          type:
            - string
            - "null"
          format: uuid
        source_language:
          type:
            - string
            - "null"
          minLength: 2
          maxLength: 16
        target_language:
          type:
            - string
            - "null"
          minLength: 2
          maxLength: 16
        source_term:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 255
        target_term:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 255
        definition:
          type:
            - string
            - "null"
        domain:
          type:
            - string
            - "null"
          maxLength: 128
        case_sensitive:
          type:
            - boolean
            - "null"
        forbidden:
          type:
            - boolean
            - "null"
        example_source:
          type:
            - string
            - "null"
        example_target:
          type:
            - string
            - "null"
    GlossarySearchRequest:
      type: object
      required:
        - source_language
        - target_language
        - source_text
      properties:
        source_language:
          type: string
          minLength: 2
          maxLength: 16
        target_language:
          type: string
          minLength: 2
          maxLength: 16
        source_text:
          type: string
          minLength: 1
        domain:
          type:
            - string
            - "null"
          maxLength: 128
        project_id:
          type:
            - string
            - "null"
          format: uuid
    GlossarySearchResponse:
      type: object
      required:
        - matches
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/GlossaryTermMatch"
    GlossaryTermMatch:
      type: object
      required:
        - term
        - start
        - end
        - matched_text
      properties:
        term:
          $ref: "#/components/schemas/GlossaryTerm"
        start:
          type: integer
          minimum: 0
        end:
          type: integer
          minimum: 0
        matched_text:
          type: string
    GlossaryImportResponse:
      type: object
      required:
        - imported_count
        - terms
      properties:
        imported_count:
          type: integer
          minimum: 0
        terms:
          type: array
          items:
            $ref: "#/components/schemas/GlossaryTerm"
    GlossaryTerm:
      type: object
      required:
        - id
        - source_language
        - target_language
        - source_term
        - target_term
        - case_sensitive
        - forbidden
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type:
            - string
            - "null"
          format: uuid
        source_language:
          type: string
        target_language:
          type: string
        source_term:
          type: string
        target_term:
          type: string
        definition:
          type:
            - string
            - "null"
        domain:
          type:
            - string
            - "null"
        case_sensitive:
          type: boolean
        forbidden:
          type: boolean
        example_source:
          type:
            - string
            - "null"
        example_target:
          type:
            - string
            - "null"
...